Correct width from int to float in line and outline

+++ src/ch/epfl/imhof/painting/Painter.java
@@ -67,12 +67,12 @@ public interface Painter {
         };
     }
     
-    static Painter line(int width, Color color, LineStyle.LineCap cap, LineStyle.LineJoin join, float[] dashingPattern) {
+    static Painter line(float width, Color color, LineStyle.LineCap cap, LineStyle.LineJoin join, float[] dashingPattern) {
         LineStyle style = new LineStyle(width, color, cap, join, dashingPattern);
         return line(style);
     }
     
-    static Painter line(int width, Color color) {
+    static Painter line(float width, Color color) {
         LineStyle style = new LineStyle(width, color);
         return line(style);
     }
@@ -90,12 +90,12 @@ public interface Painter {
         };
     }
     
-    static Painter outline(int width, Color color, LineStyle.LineCap cap, LineStyle.LineJoin join, float[] dashingPattern) {
+    static Painter outline(float width, Color color, LineStyle.LineCap cap, LineStyle.LineJoin join, float[] dashingPattern) {
         LineStyle style = new LineStyle(width, color, cap, join, dashingPattern);
         return outline(style);
     }
     
-    static Painter outline(int width, Color color) {
+    static Painter outline(float width, Color color) {
         LineStyle style = new LineStyle(width, color);
         return outline(style);
     }
